Move all loading related components into /components/loading.
