+++ client/src/view/AppBar.tsx
@@ -21,7 +21,7 @@ import React, { useState } from 'react';
 import { matchPath, useLocation } from 'react-router';
 import { LoggedInUserTutorial } from 'shared/dist/model/Tutorial';
 import { useChangeTheme } from '../components/ContextWrapper';
-import SubmitButton from '../components/forms/components/SubmitButton';
+import SubmitButton from '../components/loading/SubmitButton';
 import { getTutorialXLSX } from '../hooks/fetching/Files';
 import { useLogin } from '../hooks/LoginService';
 import { getDisplayStringForTutorial, saveBlob } from '../util/helperFunctions';
