+++ client/src/view/attendance/AttendanceManager.tsx
@@ -9,8 +9,8 @@ import { Student, StudentStatus } from 'shared/dist/model/Student';
 import { LoggedInUser, TutorInfo } from 'shared/dist/model/User';
 import CustomSelect from '../../components/CustomSelect';
 import DateOfTutorialSelection from '../../components/DateOfTutorialSelection';
-import SubmitButton from '../../components/forms/components/SubmitButton';
-import LoadingSpinner from '../../components/LoadingSpinner';
+import SubmitButton from '../../components/loading/SubmitButton';
+import LoadingSpinner from '../../components/loading/LoadingSpinner';
 import TableWithPadding from '../../components/TableWithPadding';
 import { useAxios } from '../../hooks/FetchingService';
 import { useLogin } from '../../hooks/LoginService';
