+++ client/src/view/tutorialmanagement/TutorialManagement.tsx
@@ -12,7 +12,7 @@ import TutorialForm, {
   TutorialFormState,
   TutorialFormSubmitCallback,
 } from '../../components/forms/TutorialForm';
-import LoadingSpinner from '../../components/LoadingSpinner';
+import LoadingSpinner from '../../components/loading/LoadingSpinner';
 import TableWithForm from '../../components/TableWithForm';
 import { useDialog } from '../../hooks/DialogService';
 import { useAxios } from '../../hooks/FetchingService';
