+++ client/src/view/usermanagement/UserManagement.tsx
@@ -6,9 +6,9 @@ import { Role } from 'shared/dist/model/Role';
 import { Tutorial } from 'shared/dist/model/Tutorial';
 import { CreateUserDTO, UserDTO } from 'shared/dist/model/User';
 import { getNameOfEntity } from 'shared/dist/util/helpers';
-import SubmitButton from '../../components/forms/components/SubmitButton';
+import SubmitButton from '../../components/loading/SubmitButton';
 import UserForm, { UserFormState, UserFormSubmitCallback } from '../../components/forms/UserForm';
-import LoadingSpinner from '../../components/LoadingSpinner';
+import LoadingSpinner from '../../components/loading/LoadingSpinner';
 import SnackbarWithList from '../../components/SnackbarWithList';
 import TableWithForm from '../../components/TableWithForm';
 import { useDialog } from '../../hooks/DialogService';
